Build and link for iOS and Android
Twenty-three files. The bulk of it is one recurring problem: this codebase's platform
conditionals are wrong for mobile in both directions. `__APPLE__` is too broad, so iOS
compiled macOS-only code (AppKit, IOKit paths apps cannot reach); and `_WIN32`/`__APPLE__`
pairs are too narrow, so Android fell off the end and got nothing -- sometimes an #error,
more often silence. Where a guard was wrong, the fix expresses the rule it was reaching
for rather than adding a platform to the list, which is why several hunks below mention
POSIX instead of naming Android.

iOS
  - BlueClipboard.mm: UIPasteboard branch. There is no AppKit and no NSPasteboard, and the
    API is narrower -- a single `string` property rather than typed items, so the
    availableTypeFromArray / declareTypes dance has no counterpart and a nil string is the
    only "nothing there" signal.
  - BlueSysInfo.mm: GetMachineUuid uses identifierForVendor. kIOPlatformUUIDKey needs IOKit
    APIs apps cannot reach. Per-vendor rather than per-device, and regenerated when the last
    app from that vendor is uninstalled; accepted as experiment-grade.
  - errormessage.cpp: ShowErrorMessageBox no longer blocks on a dialog and keeps its
    existing log. PyGetExeFilePids raises NotImplementedError rather than returning an empty
    list, because an empty list is a plausible *wrong* answer to "is another instance
    running".
  - blue.cpp: _Py_FatalErrorFunc is no longer hijacked. That is an improvement here, not a
    regression -- blue called kill(getpid(), SIGKILL), which produces no crash report on
    iOS, while CPython's own abort raises SIGABRT, which does.
  - BlueSysInfo.cpp: the GetPDMData gate is gone. It returned empty because RetrievePDMData
    pulls in the whole gatherer subgraph and no iOS backend existed; pdm supplies one now.
    Verified by control build: with the gate blue.so carries none of the gatherer's report
    literals, without it all three.
  - CMakeLists.txt: UIKit instead of AppKit. SystemConfiguration, Foundation and IOKit are
    all in the iOS SDK, so that is the only framework that differs.

Android
  - Two new files, following the existing one-file-per-platform convention:
    BlueSysInfoAndroid.cpp (the eleven platform methods plus the three BlueSysInfo*
    constructors) and BlueClipboardAndroid.cpp.
  - The clipboard fails rather than emulating. ClipboardManager has no NDK counterpart, and
    since Android 10 only the focused app may read the clipboard at all. An in-process
    buffer would make copy and paste appear to work while being invisible to every other
    app, which is worse than a clean failure.
  - Paths cannot be asked for without a Context, so they are derived: /proc/self/cmdline is
    the package name for the main process, and returns empty rather than a guess when the
    directory is not there. There is no user-visible Documents directory reachable without a
    permission the user must grant, so that one is app-scoped.
  - BlueSysInfoOs::Platform gained ANDROID_OS, with a chooser row alongside. This is a
    public, Python-visible enum -- additive at the end, existing values unchanged. The name
    is not ANDROID because the NDK toolchain puts -DANDROID on every compile line, so an
    enumerator of that name is macro-expanded before the compiler sees it.
  - The three .mm files are HEADER_FILE_ONLY off Apple. They are wholly guarded on
    __APPLE__ so they contribute nothing, but they are still compiled as Objective-C++
    against a plain C++ precompiled header, which fails outright on Android.
  - liblog is linked on Android, and this is a compensating fix in the wrong layer,
    knowingly. protobuf 3.6.0.1 calls __android_log_write from stubs/common.cc and its
    CMake has no Android awareness at all, so it never links liblog and every Android
    consumer inherits the unresolved symbol. The fix belongs in the protobuf port; that
    version is a v141 pin resolved from the registry's version database rather than from
    ports/protobuf, so putting it there means reconstructing a historic port tree for one
    line. Recorded rather than hidden.
  - BluePython.h/.cpp: va_list is a struct on aarch64 Linux, not a pointer, so it is passed
    by address. IBlueOS.h: bionic has no C11 Annex K errno_t. blue.cpp: flock after open,
    because O_SHLOCK and O_EXLOCK are BSD-only -- note this leaves a window between the two
    calls that macOS does not have.

Defects that are not mobile problems
  These were latent everywhere and only a new configuration exposed them.
  - BitPacker.h declared its 64-bit overloads as uint64_t and defined them as unsigned long
    long. Those are the same type on Apple and MSVC and different types wherever uint64_t is
    unsigned long, which is LP64 Linux -- so every 64-bit Pack call was ambiguous. BlueNet.cpp
    is corrected to the fixed-width spelling to match.
  - Four CCP_LOG* calls passed a runtime string as the format argument, which any build with
    -Wformat-security -Werror rejects. This is a habit rather than an accident: four
    instances in this repository alone, and turning that warning on for Windows and macOS
    would likely find more.
  - PyMemory.cpp had an #error for every platform that is neither Windows nor Apple, guarding
    a call to CCPMSize -- a carbon-core API available everywhere it builds.

+++ src/BlueNet.cpp
@@ -558,7 +558,10 @@ bool BlueNet::SendPacketToClientList( const unsigned long long* clientList,
 	{
 		TransportRepr *transport;
 		int num;
-		unsigned long long recipients[c_maxListLen];
+		// uint64_t, not unsigned long long: BitPacker declares its 64-bit overloads in the
+		// fixed-width spelling, and the two are different types wherever uint64_t is unsigned
+		// long (bionic on LP64), which makes every Pack call here ambiguous.
+		uint64_t recipients[c_maxListLen];
 	};
 
 	char headerData[sizeof(BlueNetHeader) + (sizeof(TempListNode) * 5) / 4];
@@ -1985,7 +1988,7 @@ void BlueNet::Route( BlueNetHeader* header,
 			int forkedCount = header->forkedAddresses; // preserve list but blank the header for retransmit
 			header->forkedAddresses = 0;
 
-			unsigned long long forkTargetClientID;
+			uint64_t forkTargetClientID;
 			for( int i=0; i<forkedCount; i++)
 			{
 				if ( !packer->Unpack(forkTargetClientID) )
